Move root routes to from router/index to routes/RootRouter
