+++ src/router/index.js
@@ -1,69 +1,9 @@
-import React, { useState, useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import { Route, StaticRouter } from 'react-router'
-import { Switch } from 'react-router-dom'
-import { createBrowserHistory, createMemoryHistory } from 'history'
+import React from 'react'
+import { StaticRouter } from 'react-router'
 import { ConnectedRouter } from 'connected-react-router'
-import checkLogin from 'store/actions/checkLogin'
-import getSignupState, { SignupState } from 'store/selectors/getSignupState'
-import ErrorBoundary from 'components/ErrorBoundary'
-import Loading from 'components/Loading'
-import AuthLayoutRouter from 'routes/AuthLayoutRouter'
-import PublicLayoutRouter from 'routes/PublicLayoutRouter'
-import NonAuthLayoutRouter from 'routes/NonAuthLayoutRouter'
+import { createBrowserHistory, createMemoryHistory } from 'history'
+import RootRouter from 'routes/RootRouter'
 import '../css/global/index.scss'
-
-function Router () {
-  const dispatch = useDispatch()
-  const signupState = useSelector(getSignupState)
-  const [loading, setLoading] = useState(true)
-  const isAuthorized = [SignupState.InProgress, SignupState.Complete].includes(signupState)
-
-  // This should be the only place we check for a session from the API
-  // and it should not load the router until that check is complete
-  useEffect(() => {
-    const asyncFunc = async () => {
-      setLoading(true)
-      await dispatch(checkLogin())
-      setLoading(false)
-    }
-    asyncFunc()
-  }, [dispatch, checkLogin, setLoading])
-
-  if (loading) {
-    return (
-      <Loading type='fullscreen' />
-    )
-  }
-
-  // TODO: Revisit this in terms of returnToPath
-  // On mobile we want to only store the intended URL and forward to the
-  // download app modal (which is currently on the Login component/page)
-  // Specifically we don't want any components to do any work but this,
-  // namely JoinGroup which utilizes returnToOnAuth) and may attempt
-  // to auth the user with a token and send them into sign-up.
-
-  return (
-    <ErrorBoundary>
-      {!isAuthorized && (
-        <Switch>
-          <Route path='/public' component={PublicLayoutRouter} />
-          <Route component={NonAuthLayoutRouter} />
-        </Switch>
-      )}
-      {isAuthorized && (
-        <Route
-          path={[
-            '/:context(groups)/:groupSlug/:view(events|groups|map|members|projects|settings|stream|topics)?',
-            '/:context(all|public)/:view(events|groups|map|members|projects|settings|stream|topics)?',
-            '/'
-          ]}
-          component={AuthLayoutRouter}
-        />
-      )}
-    </ErrorBoundary>
-  )
-}
 
 export const history = typeof window !== 'undefined'
   ? createBrowserHistory()
@@ -74,7 +14,7 @@ export function clientRouter () {
 
   return (
     <ConnectedRouter history={history}>
-      <Router />
+      <RootRouter />
     </ConnectedRouter>
   )
 }
@@ -83,7 +23,7 @@ export function clientRouter () {
 export function serverRouter (req, context) {
   return (
     <StaticRouter location={req.url} context={context}>
-      <Router />
+      <RootRouter />
     </StaticRouter>
   )
 }
