+++ src/routes/RootRouter/index.js
@@ -1,0 +1,3 @@
+import component from './RootRouter'
+
+export default component
